requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-docx==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
google-genai
googlesearch-python==1.2.3
python-dotenv==1.0.0
gunicorn